// be/src/modules/chat/enums/chat-error.enum.ts
export enum ChatErrorCode {
  MATCH_NOT_FOUND = 'MATCH_NOT_FOUND',
  NOT_PARTICIPANT = 'NOT_PARTICIPANT',
  MATCH_NOT_ACTIVE = 'MATCH_NOT_ACTIVE',
  MESSAGE_EMPTY = 'MESSAGE_EMPTY',
  MESSAGE_TOO_LONG = 'MESSAGE_TOO_LONG',
  IMAGE_NOT_SUPPORTED = 'IMAGE_NOT_SUPPORTED',
  INVALID_CURSOR = 'INVALID_CURSOR',
  NOT_IMPLEMENTED = 'NOT_IMPLEMENTED',
}